Operation: Cleanup
Removed the font! (Wasn't feeling it) Replaced it with ol'reliable poppins (Nice and tidy!).

The ui looked a bit too zoomed out, so i fixed this by tweaking the font sizes, the icon sizes until i had a scale I was happy with. This will provide me with a usable canvas for my news article elements.